+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - A `Debug2D` singleton to help the process of debugging in 2D by drawing. 
 - A `Spawner2D` node to automatically instantiate nodes from scenes and scripts in 2D.
 - A `GoostMath` singleton.
+- A `PolyEllipse2D` node, which generates ellipse polygon, and corresponding `GoostGeometry2D.ellipse()` method.
 - A project setting for configuring path to repository by Git plugin: `version_control/git/repository_path`.
 - An editor setting for `MixinScript` auto-switching behavior: `text_editor/files/open_first_script_on_editing_mixin_script` (disabled by default now).